Additional extra parameters for positive case AddCommand
This script is intend to check the request, which comes to SDL when the
command has all correct parameters and extra parameters: ttsChunks from another
request and fakeParams, which not from Protocol. The command should be added
to UI Commands Menu and to VR menu. All parameters are in boundary conditions.
SDL should ignore extra ttsChunks and fake params. SDL must respond with
resultCode "Success" and success:"true" value.

Related: APPLINK-31632